projects/asset_service/requirements.txt
alembic==1.0.10
aniso8601==6.0.0
attrs==19.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.7
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
flask-restplus==0.12.1
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
Mako==1.0.12
MarkupSafe==1.1.1
psycopg2-binary==2.*
PyJWT==1.7.1
pyrsistent==0.15.2
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
requests==2.22.0
setuptools==40.6.2
six==1.13.0
SQLAlchemy==1.3.4
urllib3==1.25.7
Werkzeug==0.15.5
